Added direct download to user home directory
This is a ease of use change. Some users might not have or know how to use git on CLI, so I've added a direct download. Some users might not know how to make a script world executable, so I've added that in the one-liner as well. The script is now downloaded to the users home directory, this is so the script can be called easily with 'sudo ~/autowgc' no matter the working directory. I would've preferred for the script to live in the users $PATH environment variable, however this script requires root, and editing a root users $PATH variable really shouldn't be done by anyone/anything except root.